XW-849 | server/app files renamed to dashed-case
