+++ server/app/facets/api/article-comments.js
@@ -1,8 +1,8 @@
 import {badRequest} from 'boom';
-import {PageRequest} from '../../lib/MediaWiki';
-import {getCachedWikiDomainName} from '../../lib/Utils';
+import {PageRequest} from '../../lib/mediawiki';
+import {getCachedWikiDomainName} from '../../lib/utils';
 import localSettings from '../../../config/localSettings';
-import getStatusCode from '../operations/getStatusCode';
+import getStatusCode from '../operations/get-status-code';
 
 /**
  * @typedef {Object} Comment
