+++ server/app/facets/operations/redirect-to-root.js
